Move irrelevant functions out of tapecontours
